+++ packages/FormElement/stories/examples/ListBox.js
@@ -1,0 +1,18 @@
+import React from "react";
+import ListBox from "@paprika/list-box";
+import FormElement from "../../src/FormElement";
+
+export default function ListBoxExample() {
+  const { listBoxA11yProps, formElementA11yProps } = useFormElement();
+  const optionsArray = ["Black Panther", "Wonder Woman", "Spiderman", "The Incredibles", "Thor", <span>test</span>];
+  const listboxOptions = optionsArray.map(hero => <ListBox.Option key={hero}>{hero}</ListBox.Option>);
+
+  return (
+    <FormElement formElementA11yProps={formElementA11yProps}>
+      <Label>Form Label</Label>
+      <Content>
+        <ListBox {...listBoxA11yProps}>{listboxOptions}</ListBox>
+      </Content>
+    </FormElement>
+  );
+}