+++ packages/ListBox/src/components/Option/helpers/optionState.js
@@ -25,7 +25,7 @@ function createOption({ index, child, groupLabel = null, groupId = null }) {
 }
 
 export function getDataOptions(children) {
-  if (!children) throw new Error("Listbox.Options is a required prop, please check you are passing correctly the data");
+  if (!children) throw new Error("ListBox.Options is a required prop, please check you are passing correctly the data");
   if (!children.length) return [];
 
   const options = {};