+++ packages/SideNavigation/src/SideNavigation.styles.js
@@ -1,0 +1,49 @@
+import styled from "styled-components";
+import SidePanel from "@paprika/sidepanel";
+import RawButton from "@paprika/raw-button";
+import tokens from "@paprika/tokens";
+import { spacer, fontSize } from "@paprika/stylers/lib/helpers";
+
+export const SideNavigationCollapsedWrapper = styled.div`
+  background: ${tokens.color.white};
+  border-right: ${tokens.color.blackLighten60} solid 1px;
+  padding: ${tokens.spaceSm};
+  width: ${spacer(5)};
+`;
+
+export const SidePanelContent = styled(SidePanel.Content)`
+  * {
+    box-sizing: border-box;
+  }
+
+  > nav > ul,
+  > nav > ul > li {
+    list-style: none;
+    margin: 0;
+    padding: 0;
+  }
+
+  &[data-pka-anchor="sidepanel.content"] {
+    margin: -${tokens.space};
+    padding: 0;
+  }
+`;
+
+export const Trigger = styled(RawButton)`
+  ${fontSize(3)}
+  align-items: center;
+  border-radius: ${tokens.border.radius};
+  display: flex;
+  height: ${spacer(5)};
+  justify-content: center;
+  width: ${spacer(5)};
+
+  &:hover {
+    background-color: ${tokens.color.blackLighten70};
+  }
+
+  &:focus {
+    box-shadow: ${tokens.highlight.active.withBorder.boxShadow};
+    outline: none;
+  }
+`;