+++ packages/ListBox/CHANGELOG.md
@@ -6,6 +6,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+## [1.0.0] - 2020-11-04
+
+### Changed
+
+- Moved from Listbox
+
 ## [0.9.0] - 2020-06-16
 
 ### Changed